Add APIContext
