+++ modules/context/api.go
@@ -1,0 +1,22 @@
+// Copyright 2016 The Gogs Authors. All rights reserved.
+// Use of this source code is governed by a MIT-style
+// license that can be found in the LICENSE file.
+
+package context
+
+import (
+	"gopkg.in/macaron.v1"
+)
+
+type APIContext struct {
+	*Context
+}
+
+func APIContexter() macaron.Handler {
+	return func(c *Context) {
+		ctx := &APIContext{
+			Context: c,
+		}
+		c.Map(ctx)
+	}
+}
