+++ public/less/_user.less
@@ -23,6 +23,9 @@
 
 	&.profile {
 		.ui.card {
+			#profile-avatar {
+				height: 290px;
+			}
 			.username {
 				display: block;
 			}
